Fix dropping of sensor timeouts
Previously, a sensor timeout would be interrupted by any event.
In particular, sensor activations while the lights are on made
the lights remain on indefinitely.

The cause of this issue was that the timeout was only set when
the state of the lights had changed. This is incorrect.

The solution was separating the effects caused by a transition
and the timeouts necessary in the new state into a separate function.
Importantly, the timeouts must be entirely determined by the new state,
being independent of the previous one.