Make some button input drivers fire immediately
For some button input drivers, it is best to fire
clicks immediately, even if part of a multi-click.
This leads to repeated events for the first, second,
third click and so on.

This works best when the actions of latter clicks
modify the previous actions in a compatible way.
For example: the first click turns the lights and
sensor on together, the second turns the sensor off.
Generally, we want to make sure that repeated clicks
will not generate conflicting events for home assistant
(e.g. first turn lights on, then off).

For other input drivers, when the actions they produce
are not compatible, it is more appropriate to wait until
the end of the cooldown before any click event is fired.

The previous behaviour was the latter, which is safer.
It is not always very usable, as it adds delay when
responding to buttons. This is particulary bad when
the device itself will not produce clicks within a reasonable
cooldown period, such as the on/off buttons in the IKEA Styrbar remote.