Fix Hass.WebSocket monitoring of callers
I had changed the caller table to contain not just
a list of message IDs, but also the reference to
its process monitor, so that we can unmonitor the
process once all calls are resolved.

However, I had overlooked the function unregister_handler,
which put a list of ids into the process table, causing
other parts of this module to crash.